kernel-2.6.32-504.el6
CentOS kernel, generated by rpmbuild -bp --target=noarch

+++ include/linux/swap.h
@@ -208,8 +208,8 @@ extern unsigned long totalram_pages;
 extern unsigned long totalreserve_pages;
 extern int min_free_kbytes;
 extern int extra_free_kbytes;
-extern unsigned int nr_free_buffer_pages(void);
-extern unsigned int nr_free_pagecache_pages(void);
+extern unsigned long nr_free_buffer_pages(void);
+extern unsigned long nr_free_pagecache_pages(void);
 
 /* Definition of global_page_state not available yet */
 #define nr_free_pages() global_page_state(NR_FREE_PAGES)
@@ -270,7 +270,7 @@ extern int __isolate_lru_page(struct page *page, isolate_mode_t mode, int file);
 extern unsigned long shrink_all_memory(unsigned long nr_pages);
 extern int vm_swappiness;
 extern int remove_mapping(struct address_space *mapping, struct page *page);
-extern long vm_total_pages;
+extern unsigned long vm_total_pages;
 
 #ifdef CONFIG_NUMA
 extern int zone_reclaim_mode;
